Rely on window explicitly.
See #88, #89, #92. This also reverts the behavior in #92.

We do this everywhere else, so we might as well do it here too. This
should solve the issue of undefined contexts when logging by not relying
on `this` being `window` in the factory builder.

I like this approach better because if `root` is always window anyway,
it's cleaner to call it that in the factory. `root` in the module
definition is part of the [UMD pattern](https://github.com/umdjs/umd),
and I'd like to keep it isolated.